Possibly better separation
I'm using the so-called reveal module

I also modify to make __index refer to extended type so that it can also
be used as is. My goal was to make it easy for me to see how I might
extend and add to a prototype without overriding the Sprite using
underscore.